Static analysis fixing

+++ src/main/java/innotutor/innotutor_backend/service/CardService.java
@@ -1,29 +1,6 @@
-/*
-MIT License
-
-Copyright (c) 2021 InnoTutor
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
- */
 package innotutor.innotutor_backend.service;
 
-import innotutor.innotutor_backend.DTO.card.CardDTO;
+import innotutor.innotutor_backend.dto.card.CardDTO;
 import innotutor.innotutor_backend.entity.card.Card;
 import innotutor.innotutor_backend.entity.user.Request;
 import innotutor.innotutor_backend.entity.user.User;
@@ -59,11 +36,11 @@ public class CardService {
     private final CardSessionFormatRepository cardSessionFormatRepository;
     private final CardSessionTypeRepository cardSessionTypeRepository;
 
-    public CardDTO getCardById(Long cardId, Long userId) {
-        Optional<Card> cardOptional = cardRepository.findById(cardId);
+    public CardDTO getCardById(final Long cardId, final Long userId) {
+        final Optional<Card> cardOptional = cardRepository.findById(cardId);
         if (cardOptional.isPresent()) {
-            Card card = cardOptional.get();
-            Long creatorId = new CardCreatorId(card).creatorId();
+            final Card card = cardOptional.get();
+            final Long creatorId = new CardCreatorId(card).creatorId();
             if (!card.getHidden() || userId.equals(creatorId)) {
                 return new CardDTOCreator(card, creatorId, subjectRepository).create();
             }
@@ -71,7 +48,7 @@ public class CardService {
         return null;
     }
 
-    public CardDTO postCvCard(CardDTO cardDTO) {
+    public CardDTO postCvCard(final CardDTO cardDTO) {
         return new CardCreator(cardDTO,
                 CardType.SERVICE,
                 cardRepository,
@@ -85,7 +62,7 @@ public class CardService {
                 cardSessionTypeRepository).postCard();
     }
 
-    public CardDTO postRequestCard(CardDTO cardDTO) {
+    public CardDTO postRequestCard(final CardDTO cardDTO) {
         return new CardCreator(cardDTO,
                 CardType.REQUEST,
                 cardRepository,
@@ -99,7 +76,7 @@ public class CardService {
                 cardSessionTypeRepository).postCard();
     }
 
-    public CardDTO putCvCard(CardDTO cardDTO) {
+    public CardDTO putCvCard(final CardDTO cardDTO) {
         return new CardUpdater(cardDTO,
                 CardType.SERVICE,
                 cardRepository,
@@ -113,7 +90,7 @@ public class CardService {
                 cardSessionTypeRepository).putCard();
     }
 
-    public CardDTO putRequestCard(CardDTO cardDTO) {
+    public CardDTO putRequestCard(final CardDTO cardDTO) {
         return new CardUpdater(cardDTO,
                 CardType.REQUEST,
                 cardRepository,
@@ -127,25 +104,21 @@ public class CardService {
                 cardSessionTypeRepository).putCard();
     }
 
-    public boolean deleteCardById(Long userId, Long cardId) {
-        Optional<Card> cardOptional = cardRepository.findById(cardId);
-        Optional<User> userOptional = userRepository.findById(userId);
+    public boolean deleteCardById(final Long userId, final Long cardId) {
+        final Optional<Card> cardOptional = cardRepository.findById(cardId);
+        final Optional<User> userOptional = userRepository.findById(userId);
         if (!cardOptional.isPresent() || !userOptional.isPresent()) {
             return false;
         }
-        innotutor.innotutor_backend.entity.user.Service service = serviceRepository.findByCardId(cardId);
-        if (service != null) {
-            if (service.getTutorId().equals(userId)) {
-                cardRepository.deleteById(cardId);
-                return true;
-            }
+        final innotutor.innotutor_backend.entity.user.Service service = serviceRepository.findByCardId(cardId);
+        if (service != null && service.getTutorId().equals(userId)) {
+            cardRepository.deleteById(cardId);
+            return true;
         }
-        Request request = requestRepository.findByCardId(cardId);
-        if (request != null) {
-            if (request.getStudentId().equals(userId)) {
-                cardRepository.deleteById(cardId);
-                return true;
-            }
+        final Request request = requestRepository.findByCardId(cardId);
+        if (request != null && request.getStudentId().equals(userId)) {
+            cardRepository.deleteById(cardId);
+            return true;
         }
         return false;
     }
